[Data] Removing topology scheduling fallback to guarantee liveness (#1614)
Context
---

Currently, scheduling fallback is predicated on the whole topology being
idle (ie running no tasks) for us to fallback and schedule at least 1
operator.

This is not a sensible predicate anymore with addition of incremental
HashShuffleOperator

- Shuffle has "streaming" tasks running, which are launched at the
*start* and are blocked on the input (ie not able to provide any output
until fed with substantial enough amount of input)

Changes
---

This change is evolving that pathway by removing scheduling fallback as
it's unnecessary as long as we can guarantee following:

- Each operators resource budget has to be >= a single task resource
requirement

This predicate will guarantee that so long as:

 - There are inputs (anywhere in the pipeline)
 - There enough logical resources to schedule a single task
 
At least 1 task will be scheduled.

Additionally, this change is cleaning up the code, adding full branch
coverage (in preparation to evolve the ranker).