[Data] Fix `LocationAwareBundleQueue` error that occurs with duplicate bundles (#1740)
Fixes https://anyscale1.atlassian.net/browse/DATA-1006

Bundle queues are a data structure used by `OpRuntimeMetrics` to
accurately estimate in-memory sizes, and by `MapOperator`/external
queues to prioritize bundles that exist.

Though it isn't common, you can add the same bundle multiple times to a
queue. This can happen if you do something like `ds1.union(ds2, ds2)`.

The bug was that we always decrement the total bytes of the bundle queue
when we remove a bundle, even though there could be one or more
remaining copies of that bundle in the queue.

---------

Signed-off-by: Balaji Veeramani <bveeramani@berkeley.edu>