[RLlib; Offline RL] Implement BC-IRL algorithm to enable reward model training. (#1675)
## Why are these changes needed?

Currently, RLlib includes several offline RL algorithms, but none
support learning a reward model from expert demonstrations — in other
words, none implement inverse reinforcement learning (IRL) — and then
leveraging this learned reward to train policies via reinforcement
learning.

This PR adds a state-of-the-art IRL algorithm, **BC-IRL**, which learns
a reward function such that a policy optimized with respect to this
reward aligns with the expert behavior. The algorithm performs bi-level
optimization: alternating between policy improvement and reward model
updates. **A key strength of BC-IRL** is that it produces a
**generalizable reward model**, not one limited to on-policy expert data
- enabling reward estimation beyond the expert trajectories. This
implementation supports any differentiable policy and uses PPO as the
policy learner.

>Note: BC-IRL needs an environment to run, however the environment could
use any (constant) reward b/c the reward is re-computed with the reward
model anyways. Only in evaluation you would need to have a reward that
can measure success.

>Note further: BC-IRL cannot be trained with multiple learners b/c
higher-level optimization does not work with Torch DDP. This needs to be
customized.

More specifically this PR provides the following changes:

- It provides a `BCIRLPPOConfig` class that enables the user to
configure her BC-IRL and a `BCIRLPPO` `Algorithm` class that implements
the training logic. Note, the module for BC-IRL is a `MultiRLModule`
with the `default_policy` and the `reward_module`.
- It introduces a new `AlgorithmConfig.is_online` property that signals
the `Algorithm.setup` if certain resources for online algorithms, e.g.
`EnvRunnerGroup` are needed. BC-IRL is the first algorithm in the new
API stack that comes with both online and offline experiences. This new
property enables us to implement experience mixing in the future.
- It declares a `BCIRLPPOCatalog` that configures the corresponding
`DefaultBCIRLPPORLModule` - more specfically it defines the reward
module (the policy is defined directly by our `PPOCatalog`), which comes
with three types:
       - `ACTIONS`: Inputs actions, current state, and next state,
       - `NEXT_STATE`: Inputs next_state only,
       - `CURR_NEXT_STATE`: Inputs current state, and next state.
- It uses the new `DifferentiableLearner` and `TorchMetaLearner` to
derive the `BCIRLPPOTorchMetaLearner` and the
`TorchBIRLPPODifferentiableLearner`. The former implements the
higher-level BC-style loss and the latter the lower-level PPO-style loss
(together with torch-computed (and therefore differentiable))
advantages.
- It provides a tuned example that applies BC-IRL on `CartPole-v1` and
learns consistently (a) a policy and (b) the generalizable reward model.
- It also introduces a couple of improvements for `TorchMetaLearner` and
`DifferentiableLearner`/`TorchDIfferentiableLearner` that have proven
effective in practice.

## Related issue number

<!-- For example: "Closes #1234" -->

## Checks

- [x] I've signed off every commit(by using the -s flag, i.e., `git
commit -s`) in this PR.
- [x] I've run `scripts/format.sh` to lint the changes in this PR.
- [x] I've included any doc changes needed for
https://docs.ray.io/en/master/.
- [ ] I've added any new APIs to the API Reference. For example, if I
added a
method in Tune, I've added it in `doc/source/tune/api/` under the
           corresponding `.rst` file.
- [x] I've made sure the tests are passing. Note that there might be a
few flaky tests, see the recent failures at https://flakey-tests.ray.io/
- Testing Strategy
   - [x] Unit tests
   - [ ] Release tests
   - [ ] This PR is not tested :(

---------

Signed-off-by: simonsays1980 <simon.zehnder@gmail.com>
Co-authored-by: Sven Mika <svenmika1977@gmail.com>