smart port assignment in replica (#1751)
1. Set socket reuse to false now that we are assigning unique ports to
the replica. An attempt to reuse the same port should fail.
2. added e2e tests for direct ingress
3. replica gets the right port from controller and uses that to start
the HTTP and GRPC servers
4. There is a bug in the internal error flow, hence I am skipping that
test in this PR. In a follow PR i need to investigate why 500 internal
server error causes socket to not release the port.
5. in the direct ingress flow set proxy to run on head node. This is
needed to support auto scale from zero conditions.
6. Protocol branching logic is pushed down into node_port_manager from
the controller. This makes the code cleaner.
7. include instance_id in Target object
8. controller now uses `ReplicaDetails` previously it was using
`RunningReplicaInfo`. needed to make this change because ReplicaDetails
is what contains instance_id

---------

Signed-off-by: abrar <abrar@anyscale.com>