[data] Improve actor map scheduling (#1793)
While investigating fault tolerance performance for Apple workloads, we
noticed that the scheduling loop for actor pools of 1000s of actors is
very slow (10-30s). This was tracked down to a variety of linear or
super-linear operations within the ActorPoolMapOperator, by improving
the runtime of these operations we saw the time for a scheduling loop
step drop to hundredths to tenths of a second. We will introduce some
polymorphism to allow RayTurbo only optimizations.

---------

Signed-off-by: Matthew Owen <mowen@anyscale.com>