fix user code not having access to request context in direct ingress (#1853)
fixes user code not having access to request context in user code, this
was happening after the introduction of `ReplicaResponseGenerator`, dont
know the exact cause but roughly ReplicaResponseGenerator executes each
iteration of the generator in a async task, possibly that is where the
contextvars are getting lost.
change the old way of restoring serve request context which was
applicable to queued requests with the new implementation that works for
both queued and regular requests.

---------

Signed-off-by: abrar <abrar@anyscale.com>