[serve] fix _can_accept_request behavior (#1954)
`_can_accept_request` should only allow queued requests if it's a direct
ingress request (sent directly to the http server on the replica).
Otherwise (if it's on an ingress replica with DI enabled, but the
request was sent through a handle), it should still reject requests
after `max_ongoing_requests` is exceeded.

---------

Signed-off-by: Cindy Zhang <cindyzyx9@gmail.com>