[train] Guard against edge case where elastic policy can resize to below `min_workers` (#1947)
## Summary

Fixes an edge case where the worker group can shrink below the
configured minimum workers, when the worker group is running but
allocated resources shrinks to below the minimum number of workers. Adds
a corresponding unit test.

## Where did I encounter this?

* This showed up in my elastic training benchmark prototype
https://github.com/anyscale/rayturbo/pull/1913, where I set the resource
schedule by manually reserving GPUs with the `AutoscalingCoordinator`
actor.
* To mock node preemption, I decrease logical resource availability, and
then kill a worker. In between these 2 steps, there's a small window of
time where the worker group is running healthily, but the detected
resource availability has decreased, which causes the `ScalingPolicy` to
make a downscaling decision for a running worker group.
* This is not a typical scenario, since cluster resources decreasing
while a worker group is running usually results in a worker failure,
which goes through the `ScalingPolicy` "non-running worker group"
codepath:
https://github.com/anyscale/rayturbo/blob/fb71116f4b0753f908bf385d5420a6cf3a4657a1/python/ray/anyscale/train/_internal/execution/scaling_policy/elastic.py#L83

---------

Signed-off-by: Justin Yu <justinvyu@anyscale.com>