[serve] fix test_delete_application for DI (#1893)
`test_delete_application` is failing as the `assert` statement, `assert
"Path '/' not found" in httpx.get(url).text`, is expecting a [error
message raised from the
proxy](https://github.com/anyscale/rayturbo/blob/master/python/ray/serve/_private/proxy.py#L563)
after deleting the application. For proxy mode, this works as the url
corresponds to the proxy, so the `httpx.get()` works even if the
application is deleted as the proxy is still alive. But for DI, the url
corresponds to the actual replica, not the proxy. So it errors out when
running `httpx.get(url)` as the url does not exist after the application
is deleted. In other words, proxy is making request to `:8000`, but DI
is making request to `:30000` after deleting the application.

This PR majory fixes 5 items:
1. Update get_application_urls to return client's root url,
`http://localhost:8000`, when the passed application name does not
exist. This fixes couple of tests including fix
`test_delete_application`
2. replcae all the manual string literals of urls within `test_api`,
`test_deploy`, `test_deploy_2`, `test_deploy_app`, `test_failure` with
`get_application_urls`.
3. exclude returning target groups which has the applications with
`route_prefix` set to `None` value from `get_target_groups()`
4. Imperatively created app failed to restart with controller restart as
it failed to read `route_prefix` when restarting. This is fixed to pick
up `route_prefix` correctly.
5. request_id logged in the span_log was written as string byte object.
This is fixed to be written as string after adding decode()

`test_api`, `test_deploy`, `test_deploy_app`, and `test_callback` are
all passing with DI with this PR.

This change needs to be mirrored on OSS.

---------

Signed-off-by: doyoung <doyoung@anyscale.com>