[Data] Actor tasks on non-draining nodes (#1959)
## Problem
Currently, when a node that is running a task gets pre-empted, it will
retry that task. Sometimes, this can take a while to retry because the
input objects must be reconstructed, and if there is a long chain of
operators, this will add up.

## Solution
Although we can never know which nodes will be drained, the problem
above can be mitigated by being node-status aware. For example, ray core
exposes drain apis to get the status of the node. By leveraging these
APIs, we can schedule tasks on nodes that are _not_ currently draining.

## Implementation Details
- We need a way to track draining(not drained) nodes. We do this by
adding a `DrainedNodeTracker` to centralize this information
- We need a way to track actor locations. We do this by adding a
`ActorLocationTracker` to centralize this information

## Issue
- https://anyscale1.atlassian.net/browse/CI-1112

## Tests 
- Added a test for scheduling tasks on non-draining nodes.

---------

Co-authored-by: Balaji Veeramani <balaji@anyscale.com>