[data] Parallel file reading / chunking and parallel JSONL reading (#1960)
This PR introduces the notion of file chunking and an example
implementation of it being used for line delimited JSONL files. In the
future, we want to be able to extend this to other formats like other
line delimited files such as CSVs, TSVs, etc. and other chunkable file
types such as parquet that can be chunked by row groups.

We have noticed that for jobs with large files the current file reading
story will break down especially in cases where we have spot compute /
premeptions. If we take the large files and chunk them into many smaller
chunks, this will reduce the size of the tasks and will lead to better
performance. This better performance comes from multiple avenues such as
better bin packing (if we have 100 read tasks the time to complete the
execution might be limited by the longest read task, if we instead have
10k smaller read tasks we can more fully saturate the compute) and
shorter lineage reconstruction (if a long running task fails we might
need to recompute it, if this task is much smaller it should take less
time to recompute it).

Some preliminary data shows a significant speedup for JSONL workloads
this optimization is targeting to speedup. On a fixed-size cluster with
1024 CPUs and reading 640 files:
* PR: 377s (1024 concurrent tasks, 10190 tasks in total), **170%
speedup**.
* Master: 1020s (640 concurrent tasks, 640 tasks in total)

---------

Signed-off-by: Matthew Owen <mowen@anyscale.com>