[Serve] Update app filtering from get_target_groups (#2075)
This change updates the app filtering from `get_target_groups` so the
applications that does not have `route_prefix` populated before entering
`get_target_groups` are filtered out. This prevents issues like the
followings:
-
https://buildkite.com/anyscale/rayturbo/builds/4453#0198340c-b32f-4a50-b0cc-147e7b56fedf/7-9371
-
https://buildkite.com/anyscale/rayturbo/builds/4653#01985ccf-0148-471e-8628-a0df20df8f5e/190-6704

The app statuses when `route_prefix` is not assigned yet is observed as
the following:
`ApplicationStatus.DEPLOYING`
`ApplicationStatus.DEPLOY_FAILED`
`ApplicationStatus.NOT_STARTED`

But for `ApplicationStatus.DEPLOYING`, an app can be
`ApplicationStatus.DEPLOYING` for both when `route_prefix` is populated
and not. So we cannot simply filter out the apps by the app status.

Note, `route_prefix` is populated for an app only after running
`_reconcile_build_app_task()`