[train] Add interval for get_allocated_resources calls (#2068)
1. Previously, every time the controller in [RESTARTING, INITIALIZING,
or RESCHEDULING], we send remote call to the autoscaling coordinator to
get allocated resources without intervals, this results both unnecessary
calls and "intermediate” states (11/16 GPU workers allocated).
2. Similar to send resource request, we add a 1s interval for getting
the allocated resources, if calls falls in between, it is a
NoopDecision.
3. Currently set to 1s since the controller state is before the worker
group start time, we can further tune this interval if it is too
frequent or 1s is too slow.

---------

Signed-off-by: Lehui Liu <lehui@anyscale.com>