[serve] Split up request sending and waiting for rejection response in the router (#2108)
This PR splits up `RunningReplica.send_request(...)` into two methods:
`send_request(...)` and `await_request_admission(...)` (open to name
changes for this method). Previously, `AsyncioRouter` would just call
`send_request`, which handles both dispatching the request to the
replica and getting the queue length info.

The main motivation for this split is so that we can update how queued
request and requests sent to replica metrics are updated in
`AsyncioRouter`. If `send_request(...)` is not split up, it will be
difficult to update these metrics accurately once we support UnaryUnary
gRPC call since we don't get the queue length response until the entire
request is finished.

More specifically, the queued requests metric will be incremented for
the duration of the entire request, which is wrong, and the request sent
to replica metric will be incremented after the entire request is
finished, also wrong.

the desired behavior is something like
```
# AsyncioRouter
while True:
    # increment queued request counter

    r = choose_replica()
    r.send_request()

    # decrement queued request counter
    # increment replica requests counter

    accepted = await_request_admission()  # this part can take long when we use UnaryUnary
    if accepted:
        # add callback to decrement replica requests counter
        break
```

Stack:
* __&rarr;__ https://github.com/anyscale/rayturbo/pull/2108
* https://github.com/anyscale/rayturbo/pull/2114
* https://github.com/anyscale/rayturbo/pull/2095

---------

Signed-off-by: alexyang <alexyang@anyscale.com>