[serve] Refactor how metrics are updated in router (#2114)
This PR changes how the following metrics are updated:

`ray_serve_deployment_queued_queries`:
This metric used to be incremented once the request reached the router,
and decremented once `assign_request(...)` returned. Now, since
`assign_request(...)` can take the duration of the request to finish (in
the case of Unary Unary grpc call w/ rejection), it makes more sense to
decrement the counter once the request is sent off to the replica. In
the case the request is rejected, we can re-increment the counter.

`serve_num_ongoing_requests_at_replicas`:
This metric used to be incremented once the request was accepted by the
replica, and decremented once the `ReplicaResult` finished. Since the
router may not know if the request was accepted until the entire request
finishes (in the unary unary case), we need to increment this counter
immediately after the request is sent off. We can still rely on the
`ReplicaResult` completed callback to decrement the counter, since this
is always called in the case the request is rejected or some error is
thrown.

Stack:
* https://github.com/anyscale/rayturbo/pull/2108
* __&rarr;__ https://github.com/anyscale/rayturbo/pull/2114
* https://github.com/anyscale/rayturbo/pull/2095

---------

Signed-off-by: alexyang <alexyang@anyscale.com>