[serve] Support unary unary grpc call with rejection (#2095)
To handle requests with rejection, the replica first sends back whether
the request was accepted (via initial metadata which are transmitted as
headers) then sends the actual response. Regardless if the actual
response is streaming or unary, the replica will stream back the
response. This incurs some additional cost because 3 async tasks are
created per request instead of only 1 in the UnaryUnaryCall case.

This PR adds support for handling request with rejection with a
UnaryUnaryCall to improve performance.

https://anyscale1.atlassian.net/browse/SERVE-916

Stack:
* https://github.com/anyscale/rayturbo/pull/2108
* https://github.com/anyscale/rayturbo/pull/2114
* __&rarr;__ https://github.com/anyscale/rayturbo/pull/2095

---------

Signed-off-by: alexyang <alexyang@anyscale.com>