[Data] Fix OOMs for `PandasJSONLReader` (#2130)
`PandasJSONLReader` uses an arbitrarily configured batch size (in rows)
to read data. The issue is that if each row is large, than the class
reads too much data and OOMs.

To fix this issue, this PR performs sampling to infer a good batch size.

This diff is a port of https://github.com/ray-project/ray/pull/54436,
which addressed the same issue for `PandasDatasource` on OSS.

---------

Signed-off-by: Balaji Veeramani <bveeramani@berkeley.edu>