[Serve] Enabling HTTP routes test for DI (#2100)
Mirroring Ray OSS PR: https://github.com/ray-project/ray/pull/55623

This PR updates `test_http_routes.py` to use `get_applicaiton_url()`
instead of the hard coded urls when needed. The PR also enables
`test_http_routes.py` in the DI test list.

This PR currently builds on top of the changes in this PR:
https://github.com/anyscale/rayturbo/pull/2053. The code change only
relevant to this PR can be found in the following commits:
-
[79b6abb](https://github.com/anyscale/rayturbo/pull/2100/commits/79b6abbd945f61fce357354a0c7a025bc27d01cf)
-
[6d73d3f](https://github.com/anyscale/rayturbo/pull/2100/commits/6d73d3f4f569a4187aa6d55bdb365a6dd035023b):
The comment change in this commit is updated in this commit,
[4b81cd3](https://github.com/anyscale/rayturbo/pull/2100/commits/4b81cd3de6613950f7ecd6435ddb2ab1bdbc1917).

---------

Co-authored-by: Lonnie Liu <95255098+aslonnie@users.noreply.github.com>