[Data] Add `AutoscalingCoordinator` base class so that it can be faked (#2174)
The RayTurbo autoscaler depends on the `AutoscalingCoordinator`, and the
`AutoscalingCoordinator` implementation requires an active Ray cluster.
This is problematic, because it means we can't test the RayTurbo
autoscaler without patching functions (and making assumptions about the
implementation), or starting a Ray cluster (which is slow).

To address this issue, this PR introduces an explicit interface and a
`FakeAutoscalingCoordinator` implementation.

These changes are primarily used by
https://github.com/anyscale/rayturbo/pull/2165.

Signed-off-by: Balaji Veeramani <bveeramani@berkeley.edu>