[serve] Make timeout and disconnect optional (#2179)
This PR makes request timeout and disconnect handling optional, which
improves performance by 4-8% (depending on concurrency). The improvement
comes from a reduction in async tasks that are created per request,
which crowd the event loop.

Disconnect handling can be disabled by setting the
`x-request-disconnect-disabled` http header with a value of `"?1"`. If
this header is set to `"?0"` or not provided, disconnects will still be
enabled.

Also, a custom timeout can now be provided per request via the
`x-request-timeout-seconds` http header. Timeout handling can be
disabled altogether by setting the header's value to a non-negative
value, e.g. `"0"`. If this header isn't provided, or if it's an invalid
value (`"abc"`), Serve will fallback to using the value from
`HttpOptions`.

---------

Signed-off-by: akyang-anyscale <alexyang@anyscale.com>