[Serve] Fix batch serve tracing context (#2250)
Issue: https://anyscale1.atlassian.net/browse/SERVE-1079

Context used by Open Telemetry gets fixed to the context created when
the very first request is made to the batcher due to the inherited
context value when the asyncio.Task for batch processing was created.
Due to this, our users lose trace of the emitted span of the batching
method as it always points to the trace created by the very first
request.

To fix this, we keep on track of the Open Telemetry Context when
enqueuing the requests, and set these context while processing each
batch. This way, the open telemetry context read within the batching
method corresponds to the correct context.