[Data] Fixing empty projection handling (#2253)
Context
---

This is a proper fix for the issue that was mistakenly handled
[here](https://github.com/anyscale/rayturbo/pull/2249).

There's actually no issue in Arrow's ability to combine pushed down
predicate and (empty) projection. Actual issue actually lays in the
following Arrow quirk

- When empty projection is read from Parquet file Arrow returns
`RecordBatch` wiht 0 columns and `num_rows > 0`
- However, when we concat multiple `RecordBatches` using `concat_table`
it actually collapses it to a `Table` with 0 columns that in turn only
has 0 rows. That happens b/c unlike `RecordBatch`, `Table` is not able
to have 0 columns and num_rows > 0, since it derives the size of the
table from the size of its columns.

Changes
---

To work this around we're injecting a `__stub` column into empty
projection to avoid collapsing mentioned above.