[data][train] Mid-epoch resume (8/n): Deepcopy the row ids to free the underlying blocks earlier (#2294)
In training ingest with Ray Data, we process large Arrow tables that
live in shared memory, and we extract batches of data from them using a
loop. In row-id based data checkpointing, we take a view of the arrow
table and keep it around until accumulating ~128MB, at which point we
will flush them and clear the list.

Typically, the block can be freed from shared memory as soon as the
`table_iterator` moves on to the next table.

However, because we save a bunch of views of the table, the table may
stay around for longer until those references are eventually removed
when the flush happens. We want to avoid this because the underlying
table could contain a bunch of other columns that no longer have
references (since they've been processed by the model already). The
block is only hanging around because of these remaining row id column
references.

To fix this, deepcopy the row ids of the batch in order to avoid
accumulating references to the underlying table. The table can be freed
as early as possible, which brings us back to the original behavior
without data checkpointing.

---------

Signed-off-by: Justin Yu <justinvyu@anyscale.com>