[data] Adding in env var to disable chunking (#2307)
We want to add an explicit way to turn off parallel file reading /
chunking for large files. This may be done for various reasons, e.g. if
we know that the large parquet files will always have a single row group
this can be used to prevent the creation of empty read tasks.

---------

Signed-off-by: Matthew Owen <mowen@anyscale.com>