[Data] Checkpoint Perf: Wire in Checkpoint IDs via FileManifest (#2205)
### Checkpoint Perf: Wire in Checkpoint IDs via FileManifest

**Issues with Baseline**

- Entire Checkpoint IDs block are directly passed to read tasks via
`read_files`.
- Given Checkpoint IDs block scales with number of rows in the original
dataset, this can be in GBs.
- Due to this, we observe OOMs in read tasks, especially when partial
recovery is involved. Note that for full recovery, we put in an
optimization to exclude all the row IDs, so Checkpoint IDs block size
will just be a function of file fragments.

**With this Fix**

- In File Manifest, given a file path, filter on the Checkpoint IDs
block and get the associated row IDs per row group.
- Given we use bitmaps to track checkpointed row IDs, this metadata size
is expected to be in MBs and not GBs per read task.
- This allows read tasks to work with a lot smaller Checkpoint metadata
footprint and allowing Checkpointing to scale Billions of rows and
Millions of files.

---------

Signed-off-by: Srinath Krishnamachari <srinath.krishnamachari@anyscale.com>
Co-authored-by: Alexey Kudinkin <ak@anyscale.com>