[serve] Create basic implementation of HAProxyActor (#2322)
This PR introduces a new Serve actor called HAProxyActor, which will
manage the HAProxy process running on each node in a serve cluster. The
HAProxyActor's lifecycle is managed by the Serve controller. It's
lifecycle is the same as a Serve http proxy's lifecycle.

Using the HAProxyActor is controlled by the environment variable
`ANYSCALE_RAY_SERVE_ENABLE_HA_PROXY`. If this env var is set, direct
ingress will also be enabled. Because the HAProxyActor is instantiated
similar to Serve proxy, we disable the controller setting proxy location
to only be head node such that this new actor can spawn on each node. We
also return the proxies as the targets instead of the ingress replicas
so that external ALBs will hit the HAProxy instead of the ingress
replicas.

https://anyscale1.atlassian.net/browse/SERVE-1132

---------

Signed-off-by: akyang-anyscale <alexyang@anyscale.com>