[serve] Implement haproxy draining (#2342)
This PR implements the `udpate_draining` and `is_drained` methods for
`HAProxyManager`.

When the controller tells the `HAProxyManager` to begin draining,
`HAProxyManager` should tell the `HAProxy` process to begin failing
health checks (signaling to external lbs to stop sending it requests).
It also records the time at which the health checks began failing.

When the controller tells the `HAProxyManager` to stop draining, the
`HAProxy`'s health checks should begin passing again. It clears the
drain timer.

The process of telling HAProxy to pass/fail the health checks is left as
a TODO.

The method `is_drained` will return `true` if the `HAProxy` has been
failing health checks for longer than a time threshold and if there are
no ongoing requests. Detecting whether there are ongoing requests is
left as a TODO, but will most likely involve checking the active
sessions on `HAProxy`.

---------

Signed-off-by: akyang-anyscale <alexyang@anyscale.com>