[serve] Fix autoscaling in direct ingress case (#2362)
OSS PR (https://github.com/ray-project/ray/pull/56005) changed the
`Replica` code to not rely on `should_collect_metrics()` when deciding
whether to start the metrics pusher used for autoscaling. In rayturbo,
we override this method so that metrics pusher starts when the replica
is an ingress in direct ingress mode. The oss pr caused the metrics
pusher to not start for the ingress replica, causing the direct ingress
tests to break. I renamed this method to
`should_collect_ongoing_requests()` to be clear that this only applies
to ongoing requests and not custom metrics.

This needs an OSS change, but landing in rayturbo first to unblock the
pipeline faster.

---------

Signed-off-by: akyang-anyscale <alexyang@anyscale.com>