[serve] Add LongPoll get_target_groups to HAProxyManager  (#2347)
This PR implements HAProxyManager long poll for the target groups from
the controller.

This PR also changes `get_target_groups` in the controller to return the
ingress replicas as the target groups if the caller is a proxy manager
(e.g. the `HAProxyManager`) and to return the haproxies if the caller is
anyone else (e.g. via the api). The reasoning is `HAProxyManager` needs
the ingress replicas, such that it can configure the haproxies
correctly. But other callers would want the haproxies as the targets as
they're the entrypoints into the Serve cluster.

---------

Signed-off-by: akyang-anyscale <alexyang@anyscale.com>