[serve] Disable the log buffer for HAProxyManager logs (#2467)
When throughput opt flag is turned on, we buffer logs on the request
path and flush every 1000 log lines. This also applies to HAProxyManager
logs because we use the same logger configuration as Serve proxy. But we
don't want this because HAProxyManager logs are not on the critical
request path. We disable the buffer in this PR so that the Manager's
logs are immediately reflected.

Will require an oss pr

---------

Signed-off-by: akyang-anyscale <alexyang@anyscale.com>