[serve] Abort proxy -> server connection when client disconnects (#2474)
This PR adds `option abortonclose` as the default setting in our haproxy
configuration. The
[setting](https://www.haproxy.com/documentation/haproxy-configuration-manual/latest/#4.2-option%20abortonclose)
tells haproxy to close the connection with the server when the client
disconnects, which is needed for Serve's request cancellation feature.

Although the HAProxy manual states that `abortonclose` only affects
pending or establishing connections, I've verified in Serve that the
`http.disconnect` msg is received by the Serve ingress replica even
after the connection between the haproxy and replica has been
established and the request is in progress.

---------

Signed-off-by: akyang-anyscale <alexyang@anyscale.com>