[serve] Listen for logging config updates in HAProxy (#2486)
This PR subscribes HAProxy to logging config updates via longpoll and
updates the
`test_controller_recovery:test_controller_crashes_with_logging_config`
test to work not just for Serve proxies.

The test currently sends a request to the serve proxy, and checks if the
proxy has written an access log to the proxy log file. This doesn't work
in the haproxy case because haproxy manager doesn't have access logs per
request, since the access logs are written by the ingress replica. So
this PR changes the condition to check for the proxy health check log,
which actually makes more sense in the context of the test, since it's a
debug log. This also allows us to pass the test for HAProxy as the
HAProxyManager also logs health checks.

---------

Signed-off-by: akyang-anyscale <alexyang@anyscale.com>