[serve] Fix tracing_utils test for haproxy (#2495)
This PR enables the `test_tracing_utils` test for haproxy mode. Two
tests were failing in haproxy mode:
1. `test_tracing_e2e`
2. `test_tracing_e2e_with_errors`

They were failing because the tests check that a proxy span file was
created. However, in HAProxy mode, we don't trace haproxy so we never
create this file. I modified the test to not check for this file if
we're in HAProxy mode.

Also, the tests were failing for grpc. We don't support grpc in haproxy
mode yet, so disabled grpc checks when haproxy is enabled.

~~The `test_tracing_e2e` test is currently flaky in the streaming case
because haproxy sometimes combines multiple streaming chunks from the
replica into a single chunk when sending the response to the client.~~
confirmed that this is expected behavior and clients should/can handle
it accordingly

---------

Signed-off-by: akyang-anyscale <alexyang@anyscale.com>