[serve] Enable test_streaming_response for haproxy (#2499)
This test is mostly fixed by
https://github.com/anyscale/rayturbo/pull/2474. However, it was still
flaky because haproxy may coalesce chunks from the server before sending
to the client. This is a networking optimization done by haproxy, but it
causes our tests to sometimes fail. The test checks that each received
chunk is equal to some string, e.g. `hi_1` then `hi_2`, but due to the
rechunking, the received chunk may be `hi_1hi_2`. In the context of
streaming, this is a valid response to the client as the order of the
bytes have not changed. This PR adds an artificial sleep after every
`yield` in the streaming application code to prevent the coalescing from
happening so that we can still check streaming is working and avoid the
occasional failures.

---------

Signed-off-by: akyang-anyscale <alexyang@anyscale.com>