[serve] Add a secondary proxy readiness check on serve run (#2522)
This readiness check is used for checking if the proxies are able to
route traffic to the replicas after the applications have been deployed.
this is specifically required for haproxy while scale from zero is not
supported. The issue right now is when serve.run() finishes, and we send
a request to haproxy, haproxy will return with a 503 because it has not
received the replica's target info yet.

This check will ensure serve.run() blocks until haproxy recognizes at
least one server as up. When scale from zero is properly supported (by
falling back to the serve proxy), the at least one server criteria will
be satisfied by the serve proxy. The new check is only performed on
alive proxy actors.

oss pr: https://github.com/ray-project/ray/pull/57723

---------

Signed-off-by: akyang-anyscale <alexyang@anyscale.com>