[serve] Enable direct ingress tests for haproxy (#2507)
This PR adds the `from_proxy_manager` option to `get_application_url`.
This is useful because now the caller can set that flag to guarantee the
returned url is for the ingress replica if it exists. If this flag is
not set, the returned targets are whatever's the first entrypoint into
the serve app (serve proxies in vanilla serve, ingress replicas in
direct ingress, haproxies in haproxy mode).

We need to use this flag in the `test_direct_ingress.py` file because in
some cases, we want to test haproxy and in other cases we want to test
starting from the ingress replicas. For example, we set the flag to true
for all grpc related tests so that we directly hit the ingress replicas
since grpc isn't implemented for haproxy. Another example is to check
the port assignments of the ingress replicas. However, in other cases,
we want to test functionality by hitting the haproxy, so we don't set
this flag. There are also some tests in the file that don't make sense
to run for haproxy at all, so we skip the test altogether.

The related oss changes are in this PR:
https://github.com/ray-project/ray/pull/57620

---------

Signed-off-by: akyang-anyscale <alexyang@anyscale.com>