[serve][hap] Change haproxy backend name to use app name instead of route prefix (#2565)
App names in serve are always non-empty and unique. The only case when
app name can be "" in the target groups is when the target group is for
the proxy, which we don't support in haproxy for now.

This PR also fixes test_http_routes by only flipping the
`has_received_routes` flag to true if the target groups are non-empty,
which aligns itself with the implementation in proxy_router's
`update_routes` method.

---------

Signed-off-by: akyang-anyscale <alexyang@anyscale.com>