[Data] Set default `file_extensions` for `read_parquet` (#2310)
https://github.com/ray-project/ray/pull/50092 warned that we'd be
changing the default file_extensions for Parquet from `None` to
`[parquet]`. This was the motivation:

> People often have non-Parquet files in their datasets (e.g., _SUCCESS
or stale files). However, the default for file_extensions is None, so
read_parquet tries reading the non-Parquet files. To avoid this issue,
we'll change the default file extensions to something like ["parquet"].
This PR adds a warning for that change.

This PR follows up on actually changes the default.

Corresponding OSS PR: https://github.com/ray-project/ray/pull/56481

Signed-off-by: Balaji Veeramani <bveeramani@berkeley.edu>