[Data] Fix failing Parquet ID column validation tests  (#2713)
Previously, the projection pushdown rule required two methods: one for
specifying which columns to select, and another for specifying which
columns to rename.

https://github.com/ray-project/ray/pull/58286 merged those two methods
into a single method `get_projection_map` that defines both renames and
which columns to select. As a consequence of this refactor, we now pass
in the generated ID column in `columns_rename` and `columns` to
`ParquetReader.read_files`, causing the validation tests to incorrectly
start failing.

This PR refactors the validation and its tests to fix the failures.

---------

Signed-off-by: Balaji Veeramani <bveeramani@berkeley.edu>